zrealizowany punkt 7f

+++ src/pl/lublin/wsei/java/cwiczenia/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String[] args) {
         int a = 3;
-        double b = 4.21;
+        float b = 4.21f;
         String s = "jakiś tekst";
 
         System.out.printf("a = %d, b = %.2f, s = %20s %n", a, b, s);
